Fix types in endpoints
Half of the types used in the endpoints were incorrect.
Instead of using the actual type of the response, the desired type was
used, eg:

`_post<TimeTracking>()`

when the response does actually contain a Results array. Therefore the
type needs to be:

`_post<TimeTracking[]>()`

The response is then changed to the desired type by calling the
appropriate responseHandler function, such as:

`required(response)`

or

`optionalList(response)`