Fix types and responseHandlers, rename Endpoints (#33)
* Implement absence endpoints

* Fix typo

* Move test folder into src for consistency with core

* Add absences/validate

* Fix date format in test

* Implement assertAndReply helper

* Fix assertAndReply

* Fix types in endpoints

Half of the types used in the endpoints were incorrect.
Instead of using the actual type of the response, the desired type was
used, eg:

`_post<TimeTracking>()`

when the response does actually contain a Results array. Therefore the
type needs to be:

`_post<TimeTracking[]>()`

The response is then changed to the desired type by calling the
appropriate responseHandler function, such as:

`required(response)`

or

`optionalList(response)`

* Change absence cancel param to number

* Fix misspelled absenceSDays resourename

* Fix AbsenceCancel export in index

* Cleanup responseHandlers

Removed requiredList as it is never needed, renamed optionalList to
list. Changed imports to "as responseHandlers".

* Remove fixes unrelated to absences endpoint

As suggested, I will open a separate PR

* Remove change in baseApi

* Revert "Remove change in baseApi"

This reverts commit c1869fe1cbdcbebb547d8389091346f3e833b7da.

* Append "Endpoint" to endpoint classes

* Add back explicit return types in baseApi

* Improvements from PR comments

Remove default exports to ensure consistency, eg. TimeTrackingsEndpoint instead of TimetrackingsEndpoint